switch from gym to gynmasium and upgrade to numpy 2.0

+++ nes_py/_rom.py
@@ -47,7 +47,7 @@ class ROM(object):
     @property
     def header(self):
         """Return the header of the ROM file as bytes."""
-        return self.raw_data[:16]
+        return self.raw_data[:16].astype(int)
 
     @property
     def _magic(self):
